requirements.txt
Django==2.1
pytz==2018.5
tweepy==3.6.0
djangorestframework==3.8.2
django-filter==2.0.0
